Added scripts for the Database
These scripts are initializing databases and entering the first test values.
These scripts will not affect the databases at startup in any way. Since Flyway is disabled in build.gradle.

P.S.
There is an album Mutter by Rammstein in the database. Legally, it shouldn't be there. This can have a bad effect on the protection of the course work. Therefore, this album should be deleted in the future.