Remake Role entity
Now a separate class responsible for the roles has been created.
There is also a bug - at the second start, the system crashes into an error. This is due to checking users for the presence of an administrator (as a user). Therefore, only the presence of a role in the database is checked (the admin is created here)